Moving from modules to traditional .h/.cpp files
Modules causes far too many issues with tooling not
working correcty under them. Specifically clangd
barely supports modules causing a lot of frustration.

As such I have decided to move back to traditional
source file definitions for the time being until module
support become more widespread.

+++ ShadeTech/main.cpp
@@ -1,6 +1,7 @@
-import ShadeTech;
+#include "ShadeTech.cpp"
+#include "Types.h"
 
 int main(int argc, char** argv)
 {
-    return ShadeTech::RunApplication(ShadeTech{}, argc, argv);
+    return ShadeTech::RunApplication(ShadeTech{}, (u32)argc, argv);
 }
